1) build release in debug mode is now moved to separate stage 2) multiple fixes of jenkinsfile + .jenkinsci (2)

+++ .jenkinsci/linux-post-step.groovy
@@ -4,39 +4,33 @@ def linuxPostStep() {
       if (currentBuild.currentResult == "SUCCESS" && GIT_LOCAL_BRANCH ==~ /(master|develop)/) {
         def artifacts = load ".jenkinsci/artifacts.groovy"
         def commit = env.GIT_COMMIT
-        def platform = sh(script: 'uname -m', returnStdout: true).trim()
-        filePaths = [ '/tmp/${GIT_COMMIT}-${BUILD_NUMBER}/*' ]
-        artifacts.uploadArtifacts(filePaths, sprintf('/iroha/linux/%4$s/%1$s-%2$s-%3$s', [GIT_LOCAL_BRANCH, sh(script: 'date "+%Y%m%d"', returnStdout: true).trim(), commit.substring(0,6), platform]))
+        if ( 'mac' in env.NODE_NAME ) {
+          def commit = env.GIT_COMMIT
+          filePaths = [ '\$(pwd)/build/*.tar.gz' ]
+          artifacts.uploadArtifacts(filePaths, sprintf('/iroha/macos/%1$s-%2$s-%3$s', [GIT_LOCAL_BRANCH, sh(script: 'date "+%Y%m%d"', returnStdout: true).trim(), commit.substring(0,6)]))
+        }
+        else {
+          def platform = sh(script: 'uname -m', returnStdout: true).trim()
+          filePaths = [ '/tmp/${GIT_COMMIT}-${BUILD_NUMBER}/*' ]
+          artifacts.uploadArtifacts(filePaths, sprintf('/iroha/linux/%4$s/%1$s-%2$s-%3$s', [GIT_LOCAL_BRANCH, sh(script: 'date "+%Y%m%d"', returnStdout: true).trim(), commit.substring(0,6), platform]))
+        }
+        
       }
     }
     finally {
-      if (env.BUILD_TYPE == 'Debug') {
+      if ( 'mac' in env.NODE_NAME )
+      {
+        sh """
+          pg_ctl -D /var/jenkins/${GIT_COMMIT}-${BUILD_NUMBER}/ stop && \
+          rm -rf /var/jenkins/${GIT_COMMIT}-${BUILD_NUMBER}/
+        """
+      }
+      else {
         def cleanup = load ".jenkinsci/docker-cleanup.groovy"
         cleanup.doDockerCleanup()
       }
       cleanWs()
     }
-  }
-}
-
-def macPostStep() {
-  timeout(time: 600, unit: "SECONDS") {
-    try {
-      if (currentBuild.currentResult == "SUCCESS" && GIT_LOCAL_BRANCH ==~ /(master|develop)/) {
-        def artifacts = load ".jenkinsci/artifacts.groovy"
-        def commit = env.GIT_COMMIT
-        filePaths = [ '\$(pwd)/build/*.tar.gz' ]
-        artifacts.uploadArtifacts(filePaths, sprintf('/iroha/macos/%1$s-%2$s-%3$s', [GIT_LOCAL_BRANCH, sh(script: 'date "+%Y%m%d"', returnStdout: true).trim(), commit.substring(0,6)]))                        
-      }
-    }
-    finally {
-      cleanWs()
-      sh """
-        pg_ctl -D /var/jenkins/${GIT_COMMIT}-${BUILD_NUMBER}/ stop && \
-        rm -rf /var/jenkins/${GIT_COMMIT}-${BUILD_NUMBER}/
-      """
-    }
-  }
 }
 
 return this
